+++ src/app/(frontend)/(protected)/profile/account/_components/Container.tsx
@@ -1,0 +1,3 @@
+export const Container = ({ children }: { children: React.ReactNode }) => {
+    return <div className="bg-muted-background flex flex-col gap-6 rounded-md p-6">{children}</div>;
+};
